+++ example-junit5/src/test/java/com.example/AccountTest.java
@@ -18,8 +18,7 @@ import org.junit.jupiter.api.TestMethodOrder;
 public class AccountTest {
   @Test
   public void homepage() {
-    expectSelfie(get("/").body().asString())
-        .toBe(
+    expectSelfie(get("/").body().asString()).toBe(
             """
 <html><body>
 \s <h1>Please login</h1>
@@ -32,46 +31,59 @@ public class AccountTest {
 
   @Test
   public void T01_not_logged_in() {
-    expectSelfie(get("/"))
-        .toBe(
+    expectSelfie(get("/")).toBe(
             """
-<html><body>
+<html>
+\s<head></head>
+\s<body>
 \s <h1>Please login</h1>
 \s <form action="/login" method="post">
-\s   <input type="text" name="email" placeholder="email">
-\s   <input type="submit" value="login">
+\s  <input type="text" name="email" placeholder="email"> <input type="submit" value="login">
 \s </form>
-</body></html>
+\s</body>
+</html>
+╔═ [md] ═╗
+Please login
 ╔═ [statusLine] ═╗
 HTTP/1.1 200 OK""");
   }
 
   @Test
   public void T02_login(Jooby app) {
-    expectSelfie(given().param("email", "user@domain.com").post("/login"))
-        .toBe(
+    expectSelfie(given().param("email", "user@domain.com").post("/login")).toBe(
             """
-<html><body><h1>Email sent!</h1><p>Check your email for your login link.</p></body></html>
+<html>
+\s<head></head>
+\s<body>
+\s <h1>Email sent!</h1>
+\s <p>Check your email for your login link.</p>
+\s</body>
+</html>
+╔═ [md] ═╗
+Email sent!
+
+Check your email for your login link.
 ╔═ [statusLine] ═╗
 HTTP/1.1 200 OK""");
     var email = EmailDev.waitForIncoming(app);
-    expectSelfie(email)
-        .toBe(
+    expectSelfie(email).toBe(
             """
-Click <a href="http://localhost:8911/login-confirm/erjchFY=">here</a> to login.
+Click <a href="https://www.example.com/login-confirm/erjchFY=">here</a> to login.
+╔═ [md] ═╗
+Click [here](https://www.example.com/login-confirm/erjchFY=) to login.
 ╔═ [metadata] ═╗
 subject=Login to example.com
-to=user@domain.com from=team@example.com""");
+to=user@domain.com
+from=team@example.com""");
   }
 
   @Test
   public void T03_login_confirm() {
-    expectSelfie(headersToString(get("/login-confirm/erjchFY=")))
-        .toBe(
+    expectSelfie(headersToString(get("/login-confirm/erjchFY="))).toBe(
             """
 HTTP/1.1 302 Found
 content-type=text/html;charset=UTF-8
-set-cookie=login=user@domain.com;Path=/
+set-cookie=login=user@domain.com|JclThw==;Path=/
 location=/""");
   }